Remove plan9 rename function.
This function was misleading because it was actually *not* atomic
(meaning it should have never been used at all where atomicity was
required).  This will break builds on Plan 9 but I consider this fine
to do since the soon-to-be-integrated waddrmgr package relies on bolt
as a database and also does not compile for Plan 9.